edit reply migration to fix foreign key error

+++ database/migrations/2020_05_12_104745_create_replies_table.php
@@ -17,7 +17,7 @@ class CreateRepliesTable extends Migration
             $table->bigIncrements('id');
             $table->text('body');
 
-            $table->integer('question_id')->unsigned();
+            $table->unsignedBigInteger('question_id')->unsigned(); // It's hasnt To be Integer for some reasons, acttully don't know why ;)
             $table->integer('user_id');
 
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
